+++ CHANGES.md
@@ -3,8 +3,12 @@ oneko
 
 # Change log
 
-## V2.00, January 2019
+## V2.0.0, January 2019
 * Restructured Werner Randelshofer's v.1.0
 * Added support for multiple monitors.
 * Added new "window mode" -- click on cat to move it into/outof a window.
 * Cleaned up animation
+
+## V2.0.1, February 2019
+* Added mouse offset
+* Fixed defect when un-miminizing window mode.